Added Episode class which saves an episode to disk. Episode file is shared/last_episode.jsonl In Json Lines format: one frame per line. A frame exists of changes to creatures and changes to the grid.
Should I have created a separate branche for this?

Anyhow, next step is loading and playing the episode.
Even though it is a json Lines format which can be read and written line
by line I think that, considering our planned move to multithreading,
the file should be saved and written as a unity operation that cannot be
interupted. We cannot have the file being changed halfway playback.